Make sub total plots

make_sub_tot_plts() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sub total plots. The function returns Composite (a plot).

Usage

make_sub_tot_plts(
  data_tb,
  col_nms_chr,
  heights_int,
  plot_rows_cols_pair_int,
  add_legend_1L_lgl = T,
  axis_text_sclg_1L_dbl = 1,
  axis_title_sclg_1L_dbl = 1,
  legend_sclg_1L_dbl = 1,
  make_log_log_tfmn_1L_lgl = F,
  round_var_nm_1L_chr = "round",
  x_labels_chr = character(0),
  y_label_1L_chr = "Percentage"
)

Arguments

data_tb

Data (a tibble)

col_nms_chr

Column names (a character vector)

heights_int

Heights (an integer vector)

plot_rows_cols_pair_int

Plot rows columns pair (an integer vector)

add_legend_1L_lgl

Add legend (a logical vector of length one), Default: T

axis_text_sclg_1L_dbl

Axis text scaling (a double vector of length one), Default: 1

axis_title_sclg_1L_dbl

Axis title scaling (a double vector of length one), Default: 1

legend_sclg_1L_dbl

Legend scaling (a double vector of length one), Default: 1

make_log_log_tfmn_1L_lgl

Make log log transformation (a logical vector of length one), Default: F

round_var_nm_1L_chr

Round variable name (a character vector of length one), Default: 'round'

x_labels_chr

X labels (a character vector), Default: character(0)

y_label_1L_chr

Y label (a character vector of length one), Default: 'Percentage'

Value

Composite (a plot)