Make subtotal plot

make_subtotal_plt() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make subtotal plot. The function returns Subtotal (a plot).

make_subtotal_plt(
  data_tb,
  var_nm_1L_chr,
  round_var_nm_1L_chr = "round",
  x_label_1L_chr,
  y_label_1L_chr = "Percentage",
  y_scale_scl_fn = scales::percent,
  use_bw_theme_1L_lgl = T,
  legend_position_1L_chr = "none",
  label_fill_1L_chr = NULL
)

Arguments

data_tb

Data (a tibble)

var_nm_1L_chr

Variable name (a character vector of length one)

round_var_nm_1L_chr

Round variable name (a character vector of length one), Default: 'round'

x_label_1L_chr

X label (a character vector of length one)

y_label_1L_chr

Y label (a character vector of length one), Default: 'Percentage'

y_scale_scl_fn

Y scale scale (a function), Default: scales::percent

use_bw_theme_1L_lgl

Use black and white theme (a logical vector of length one), Default: T

legend_position_1L_chr

Legend position (a character vector of length one), Default: 'none'

label_fill_1L_chr

Label fill (a character vector of length one), Default: NULL

Value

Subtotal (a plot)