Make variable by round plot

make_var_by_round_plt() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make variable by round plot. The function returns Variable by round (a plot).

Usage

make_var_by_round_plt(
  data_tb,
  var_nm_1L_chr,
  x_label_1L_chr,
  label_fill_1L_chr = "Data collection",
  legend_sclg_1L_dbl = 1,
  axis_text_sclg_1L_dbl = 1,
  axis_title_sclg_1L_dbl = 1,
  round_var_nm_1L_chr = "round",
  y_label_1L_chr = "Percentage",
  y_scale_scl_fn = scales::percent
)

Arguments

data_tb

Data (a tibble)

var_nm_1L_chr

Variable name (a character vector of length one)

x_label_1L_chr

X label (a character vector of length one)

label_fill_1L_chr

Label fill (a character vector of length one), Default: 'Data collection'

legend_sclg_1L_dbl

Legend scaling (a double vector of length one), Default: 1

axis_text_sclg_1L_dbl

Axis text scaling (a double vector of length one), Default: 1

axis_title_sclg_1L_dbl

Axis title scaling (a double vector of length one), Default: 1

round_var_nm_1L_chr

Round variable name (a character vector of length one), Default: 'round'

y_label_1L_chr

Y label (a character vector of length one), Default: 'Percentage'

y_scale_scl_fn

Y scale scale (a function), Default: scales::percent

Value

Variable by round (a plot)