youthvars: Describe and Validate Youth Mental Health Economic Datasets

Modules from the readyforwhatsnext youth mental health economic model (https://readyforwhatsnext.org/) that can be used to describe and quality assure individual human record datasets. The main motivation for this package is to facilitate automated data integrity checks, ensure that methods are applied to the appropriate data structures and streamline reporting of descriptive statistics. This development version of the youthvars package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

Modules from the readyforwhatsnext youth mental health economic model (https://readyforwhatsnext.org/) that can be used to describe and quality assure individual human record datasets. The main motivation for this package is to facilitate automated data integrity checks, ensure that methods are applied to the appropriate data structures and streamline reporting of descriptive statistics. This development version of the youthvars package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

See also

Author

Maintainer: Matthew Hamilton matthew.hamilton1@monash.edu (ORCID) [copyright holder]

Authors:

Caroline Gao caroline.gao@orygen.org.au (ORCID)

Other contributors:

Orygen [copyright holder, funder]

Headspace [funder]

National Health and Medical Research Council [funder]