youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)

Create a new valid instance of the youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)

youthvars_sofas(x = make_pt_youthvars_sofas())

make_pt_youthvars_sofas()

is_youthvars_sofas(x)

Arguments

x

An object of any type

Value

A validated instance of the youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS) A prototype for youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS) A logical value, TRUE if a valid instance of the youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)

Details

youthvars S3 class for Social and Occupational Functioning Assessment Scale (SOFAS)