Synthetic population replication dataset

A purely synthetic dataset, representative of the original study data, that can be used for replication runs of package algorithms.

replication_popl_tb

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1711 rows and 43 columns.

Details

A tibble

fkClientID

unique client identifier

round

round of data collection

d_interview_date

date of data collection

d_age

age

d_gender

gender

d_sex_birth_s

sex at birth

d_sexual_ori_s

sexual orientation

d_ATSI

Aboriginal or Torres Strait Islander

d_country_bir_s

country Of birth

d_english_home

speaks English at home

d_english_native

native English speaker

d_studying_working

education and employment status

d_relation_s

relationship status

s_centre

service centre name

c_p_diag_s

primary diagnosis

c_clinical_staging_s

clinical stage

k6_total

Kessler Psychological Distress Scale (6 Dimension)

phq9_total

Patient Health Questionnaire

bads_total

Behavioural Activation for Depression Scale

gad7_total

Generalised Anxiety Disorder Scale

oasis_total

Overall Anxiety Severity and Impairment Scale

scared_total

Screen for Child Anxiety Related Disorders

c_sofas

Social and Occupational Functioning Assessment Scale

aqol6d_q1

Assessment of Quality of Life (6 Dimension) question 1

aqol6d_q2

Assessment of Quality of Life (6 Dimension) question 2

aqol6d_q3

Assessment of Quality of Life (6 Dimension) question 3

aqol6d_q4

Assessment of Quality of Life (6 Dimension) question 4

aqol6d_q5

Assessment of Quality of Life (6 Dimension) question 5

aqol6d_q6

Assessment of Quality of Life (6 Dimension) question 6

aqol6d_q7

Assessment of Quality of Life (6 Dimension) question 7

aqol6d_q8

Assessment of Quality of Life (6 Dimension) question 8

aqol6d_q9

Assessment of Quality of Life (6 Dimension) question 9

aqol6d_q10

Assessment of Quality of Life (6 Dimension) question 10

aqol6d_q11

Assessment of Quality of Life (6 Dimension) question 11

aqol6d_q12

Assessment of Quality of Life (6 Dimension) question 12

aqol6d_q13

Assessment of Quality of Life (6 Dimension) question 13

aqol6d_q14

Assessment of Quality of Life (6 Dimension) question 14

aqol6d_q15

Assessment of Quality of Life (6 Dimension) question 15

aqol6d_q16

Assessment of Quality of Life (6 Dimension) question 16

aqol6d_q17

Assessment of Quality of Life (6 Dimension) question 17

aqol6d_q18

Assessment of Quality of Life (6 Dimension) question 18

aqol6d_q19

Assessment of Quality of Life (6 Dimension) question 19

aqol6d_q20

Assessment of Quality of Life (6 Dimension) question 20