Write descriptive plots

write_descv_plots() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write descriptive plots. The function returns Descriptive plots paths (a list).

write_descv_plots(
  data_tb,
  ds_descvs_ls,
  descv_outp_dir_1L_chr,
  lbl_nms_chr = c("Household tasks", "Getting around", "Morbility", "Self care",
    "Enjoy close rels", "Family rels", "Community involvement", "Despair", "Worry",
    "Sad", "Agitated", "Energy level", "Control", "Coping", "Frequency of pain",
    "Degree of pain", "Pain interference", "Vision", "Hearing", "Communication"),
  maui_domains_pfxs_1L_chr = "vD",
  item_plots_params_ls = list(plot_rows_cols_pair_int = c(5L, 4L), heights_int = c(10L,
    1L), width_1L_dbl = 9),
  dim_plots_params_ls = list(plot_rows_cols_pair_int = c(3L, 2L), heights_int = c(10L,
    1L), width_1L_dbl = 8),
  utl_by_rnd_plots_params_ls = list(width_1L_dbl = 6, height_1L_dbl = 4),
  combined_plot_params_ls = list(nrow_1L_int = 2L, rel_heights_dbl = c(4, 10),
    scale_dbl = c(0.9, 0.9), base_height_dbl = 10)
)

Arguments

data_tb
Data (a tibble)

ds_descvs_ls
Dataset descriptives (a list)

descv_outp_dir_1L_chr
Descriptive output directory (a character vector of length one)

lbl_nms_chr
Label names (a character vector), Default: c("Household tasks", "Getting around", "Morbility", "Self care", "Enjoy close rels", "Family rels", "Community involvement", "Despair", "Worry", "Sad", "Agitated", "Energy level", "Control", "Coping", "Frequency of pain", "Degree of pain", "Pain interference", "Vision", "Hearing", "Communication")

maui_domains_pfxs_1L_chr
Multi-attribute utility instrument domains prefixes (a character vector of length one), Default: 'vD'

item_plots_params_ls
Item plots parameters (a list), Default: list(plot_rows_cols_pair_int = c(5L, 4L), heights_int = c(10L, 1L), width_1L_dbl = 9)

dim_plots_params_ls
Dimension plots parameters (a list), Default: list(plot_rows_cols_pair_int = c(3L, 2L), heights_int = c(10L, 1L), width_1L_dbl = 8)

utl_by_rnd_plots_params_ls
Utility by round plots parameters (a list), Default: list(width_1L_dbl = 6, height_1L_dbl = 4)

combined_plot_params_ls
Combined plot parameters (a list), Default: list(nrow_1L_int = 2L, rel_heights_dbl = c(4, 10), scale_dbl = c(0.9, 0.9), base_height_dbl = 10)

Value

Descriptive plots paths (a list)